Placed AddressBook.java and NetworkDataSource.java into the networkExercise package
